attributes moved to Parquet namespace
